Fix and feat: Fix error in ProjectShow view and add some buttons
