+++ src/components/project/projectList/ProjectList.jsx
@@ -4,10 +4,10 @@ import Button from 'react-bootstrap/Button';
 // import { Link } from 'react-router-dom';
 
 
-export function ProjectList({id, type, title, description, daysAgo}) {
+export function ProjectList({id, topic, title, description, company, date }) {
     return (
         <Card className="text-center">
-        <Card.Header> Proyecto de {type} </Card.Header>
+        <Card.Header> Proyecto de {topic} </Card.Header>
         <Card.Body>
             <Card.Title>{title}</Card.Title>
             <Card.Text>
@@ -16,7 +16,7 @@ export function ProjectList({id, type, title, description, daysAgo}) {
             <Button variant="primary" href={`/projects/${id}`}>Ver en más detalle</Button>
             {/* <Link to={`/projects/${id}`}>Ver en más detalle</Link> */}
         </Card.Body>
-        <Card.Footer className="text-muted">Hace {daysAgo} días</Card.Footer>
+        <Card.Footer className="text-muted">{company}</Card.Footer>
         </Card>
     )
 }
