+++ src/components/project/projectShow/financingInfo/FinancingInfo.css
@@ -1,6 +1,6 @@
 .infoOfFinancing {
     width: 100%;
-    height: 10%;
+    height: 15%;
     box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
     margin-bottom: 20px;
     border-radius: 8px;
