+++ src/components/project/registerProject/RegisterProjectForm.css
@@ -1,4 +1,4 @@
-.card-profile {
+.card-profile-form {
     width: 100%;
     height: 50%;
     background-color: rgba(131, 191, 226, 0.87);
